resolve some comments
